GSTRA/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1
langchain-ollama==0.2.0
langchain-chroma==0.1.4
chromadb==0.5.11
sentence-transformers==3.1.1
pypdf==5.0.1
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
python-dotenv==1.0.1
orjson==3.10.7
